Test for query data changing
